Javascript Hooks
This commit introduces the concept of hook to PREvant that can be used to
manipulate the deployment before handing it over to the actual infrastructure
layer. With hooks in place you are able to add custom logik, such as
reconfiguring services to use a different database configuration.

See README.md files for example usage.

Further, the commit fixes some clippy hints.
